more refactoring for code consistency
